refactor(server): replace controller method with service method
