+++ packages/server/app/Controllers/Ws/MtvRoomsWsController.ts
@@ -1,7 +1,4 @@
-import { randomUUID } from 'crypto';
 import {
-    MtvCreateWorkflowResponse,
-    MtvRoomClientToServerCreateArgs,
     MtvRoomGetRoomConstraintDetailsCallbackArgs,
     MtvRoomSummary,
     MtvRoomUpdateControlAndDelegationPermissionArgs,
@@ -12,13 +9,10 @@ import {
 import MtvRoom from 'App/Models/MtvRoom';
 import MtvRoomInvitation from 'App/Models/MtvRoomInvitation';
 import User from 'App/Models/User';
-import GeocodingService from 'App/Services/GeocodingService';
 import SocketLifecycle from 'App/Services/SocketLifecycle';
 import UserService from 'App/Services/UserService';
 import { isPointWithinRadius } from 'geolib';
-import MtvServerToTemporalController, {
-    MtvRoomPhysicalAndTimeConstraintsWithCoords,
-} from '../Http/Temporal/MtvServerToTemporalController';
+import MtvServerToTemporalController from '../Http/Temporal/MtvServerToTemporalController';
 
 interface UserID {
     userID: string;
@@ -40,10 +34,6 @@ interface RunID {
     runID: string;
 }
 
-interface OnCreateArgs extends DeviceID {
-    params: MtvRoomClientToServerCreateArgs;
-    user: User;
-}
 interface OnJoinArgs extends UserID, DeviceID {
     joiningRoom: MtvRoom;
 }
@@ -113,138 +103,6 @@ interface OnGetRoomConstraintsDetailsArgs {
 }
 
 export default class MtvRoomsWsController {
-    public static async onCreate({
-        params,
-        user,
-        deviceID,
-    }: OnCreateArgs): Promise<MtvCreateWorkflowResponse> {
-        let physicalAndTimeConstraintsWithCoords:
-            | MtvRoomPhysicalAndTimeConstraintsWithCoords
-            | undefined;
-        let creatorFitsPositionConstraint: boolean | undefined;
-        const roomID = randomUUID();
-        const room = new MtvRoom();
-        let roomHasBeenSaved = false;
-        const userID = user.uuid;
-        console.log(`USER ${userID} MTV_CREATE_ROOM ${roomID}`);
-
-        if (
-            params.hasPhysicalAndTimeConstraints &&
-            params.physicalAndTimeConstraints !== undefined
-        ) {
-            const coords = await GeocodingService.getCoordsFromAddress(
-                params.physicalAndTimeConstraints.physicalConstraintPlaceID,
-            );
-            physicalAndTimeConstraintsWithCoords = {
-                ...params.physicalAndTimeConstraints,
-                physicalConstraintPosition: coords,
-            };
-
-            //Checking for last known user position, and checking if it fits
-            //With currently creating room position constraint to allow the creator
-            //to vote for it's initial track
-            creatorFitsPositionConstraint =
-                await this.checkUserDevicesPositionIfRoomHasPositionConstraints(
-                    {
-                        user,
-                        roomConstraintInformation: {
-                            constraintLat:
-                                physicalAndTimeConstraintsWithCoords
-                                    .physicalConstraintPosition.lat,
-                            constraintLng:
-                                physicalAndTimeConstraintsWithCoords
-                                    .physicalConstraintPosition.lng,
-                            constraintRadius:
-                                physicalAndTimeConstraintsWithCoords.physicalConstraintRadius,
-                            hasPositionAndTimeConstraints:
-                                params.hasPhysicalAndTimeConstraints,
-                        },
-                        //The workflow isn't already created
-                        persistToTemporalRequiredInformation: undefined,
-                    },
-                );
-            console.log({ creatorFitsPositionConstraint });
-        }
-
-        /**
-         * We need to create the socket-io room before the workflow
-         * because we don't know if temporal will answer faster via the acknowledge
-         * mtv room creation activity than adonis will execute this function
-         */
-        const roomCreator = await User.findOrFail(userID);
-        await UserService.joinEveryUserDevicesToRoom(roomCreator, roomID);
-
-        /**
-         * Check for room name duplication
-         * If room name is found in db
-         * Just add creator nickame after the room name
-         */
-        const roomWithCreatedRoomName = await MtvRoom.findBy(
-            'name',
-            params.name,
-        );
-        const roomNameIsAlreadyTaken = roomWithCreatedRoomName !== null;
-        if (roomNameIsAlreadyTaken) {
-            params.name = `${params.name} (${roomCreator.nickname})`;
-        }
-        ///
-
-        try {
-            const temporalResponse =
-                await MtvServerToTemporalController.createMtvWorkflow({
-                    workflowID: roomID,
-                    userID: userID,
-                    deviceID,
-                    params: {
-                        ...params,
-                        physicalAndTimeConstraints:
-                            physicalAndTimeConstraintsWithCoords,
-                        creatorFitsPositionConstraint,
-                    },
-                });
-
-            room.merge({
-                uuid: roomID,
-                runID: temporalResponse.runID,
-                name: params.name,
-                //By setting this field lucid will manage the BelongsTo relationship
-                creatorID: userID,
-                isOpen: params.isOpen,
-            });
-
-            if (
-                params.hasPhysicalAndTimeConstraints &&
-                params.physicalAndTimeConstraints !== undefined &&
-                physicalAndTimeConstraintsWithCoords
-            ) {
-                room.merge({
-                    hasPositionAndTimeConstraints:
-                        params.hasPhysicalAndTimeConstraints,
-                    constraintRadius:
-                        physicalAndTimeConstraintsWithCoords.physicalConstraintRadius,
-                    constraintLat:
-                        physicalAndTimeConstraintsWithCoords
-                            .physicalConstraintPosition.lat,
-                    constraintLng:
-                        physicalAndTimeConstraintsWithCoords
-                            .physicalConstraintPosition.lng,
-                });
-            }
-            await room.save();
-            roomHasBeenSaved = true;
-
-            await roomCreator.merge({ mtvRoomID: roomID }).save();
-            await room.related('members').save(roomCreator);
-            console.log('created room ' + roomID);
-            return temporalResponse;
-        } catch (error) {
-            await SocketLifecycle.deleteSocketIoRoom(roomID);
-            if (roomHasBeenSaved) await room.delete();
-
-            throw error;
-        }
-    }
-
     public static async onJoin({
         userID,
         joiningRoom: { runID, uuid: roomID, isOpen, creatorID },
